feat: add support for jest-circus
