+++ package.json
@@ -52,7 +52,6 @@
     "@babel/core": "^7.14.6",
     "@babel/preset-env": "^7.14.7",
     "@babel/preset-typescript": "^7.14.5",
-    "@types/jasmine": "^3.7.4",
     "@types/jest": "^26.0.23",
     "@types/lodash": "^4.14.106",
     "@types/node": "^14.17.0",
